+++ internal/app/app.go
@@ -9,12 +9,12 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/zongqichen/cfs/internal/config"
-	"github.com/zongqichen/cfs/internal/envvar"
-	"github.com/zongqichen/cfs/internal/executable"
-	"github.com/zongqichen/cfs/internal/lock"
-	"github.com/zongqichen/cfs/internal/runner"
-	"github.com/zongqichen/cfs/internal/updatecheck"
+	"github.com/zongqichen/cloud-foundry-cli-contexts/internal/config"
+	"github.com/zongqichen/cloud-foundry-cli-contexts/internal/envvar"
+	"github.com/zongqichen/cloud-foundry-cli-contexts/internal/executable"
+	"github.com/zongqichen/cloud-foundry-cli-contexts/internal/lock"
+	"github.com/zongqichen/cloud-foundry-cli-contexts/internal/runner"
+	"github.com/zongqichen/cloud-foundry-cli-contexts/internal/updatecheck"
 	"golang.org/x/term"
 )