+++ internal/app/commands.go
@@ -7,8 +7,8 @@ import (
 	"io"
 	"slices"
 
-	"github.com/zongqichen/cfs/internal/config"
-	"github.com/zongqichen/cfs/internal/envvar"
+	"github.com/zongqichen/cloud-foundry-cli-contexts/internal/config"
+	"github.com/zongqichen/cloud-foundry-cli-contexts/internal/envvar"
 )
 
 const (