+++ internal/cfhome/cfhome.go
@@ -10,8 +10,8 @@ import (
 	"runtime"
 	"strings"
 
-	"github.com/zongqichen/cfs/internal/pathutil"
-	"github.com/zongqichen/cfs/internal/securefs"
+	"github.com/zongqichen/cloud-foundry-cli-contexts/internal/pathutil"
+	"github.com/zongqichen/cloud-foundry-cli-contexts/internal/securefs"
 )
 
 const (